Re-write dynamic TV show HTML generation scripts
Several things here. The previous script was too rigid and could not adapt to unlimited amount of rules. It was all hardcoded using FOR loops. Bad practice, slow execution. Using CSV files for rules, we can make use of ForEach functions that are faster and leaner. The result is a truly dynamically generated HTML table of rules to enjoy. CSV files are easier to maintain and generate as well. ShowChoice has been deleted, it's no longer needed. This entire approach also removes the reliance on cookies. Big win all around.